refactor: refatorar pasta de testes, deletar common desnecessario e colocar nomes mais descritivos para os testes
